Python backdoor
This is a simple educational python backdoor program example.  What this program do is to connect to a port of a host IPv4 address, once connected, it will recieve commands from the connected host, then execute the commands in the victims computer, and return the results to the host. to be a host, you should execute 'nc -l "port" ' command and make sure the backdoor's target IP is yours.